requirements.txt
absl-py==1.4.0
astunparse==1.6.3
cachetools==5.3.1
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.2.0
contourpy==1.1.0
cycler==0.10.0
editdistance==0.6.2
efficientnet==1.0.0
essential-generators==1.0
flatbuffers==23.5.26
fonttools==4.42.0
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.57.0
h5py==3.9.0
idna==2.10
imageio==2.31.1
imgaug==0.4.0
importlib-metadata==6.8.0
importlib-resources==6.0.1
keras==2.13.1
Keras-Applications==1.0.8
keras-ocr==0.9.2
kiwisolver==1.4.4
lazy_loader==0.3
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.76
opt-einsum==3.3.0
packaging==23.1
Pillow==9.5.0
protobuf==4.24.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyclipper==1.3.0.post4
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==1.0.0
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
roboflow==1.1.2
rsa==4.9
scikit-image==0.21.0
scipy==1.11.1
shapely==2.0.1
six==1.16.0
supervision==0.13.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-macos==2.13.0
termcolor==2.3.0
tifffile==2023.8.12
tqdm==4.66.1
typing_extensions==4.5.0
urllib3==1.26.16
validators==0.21.2
Werkzeug==2.3.6
wget==3.2
wrapt==1.15.0
zipp==3.16.2
